moving launch to a folder
