+++ src/app.css
@@ -3,7 +3,7 @@ button {
 }
 
 body {
-  /*background-color: rgb(231, 235, 240);*/
+  background-color: rgb(231, 235, 240);
 }
 
 .p-listbox-item:focus {
