+++ src/containers/subjects-launch/SubjectWGSTNLaunch/aws.ts
